+++ feature/detail/src/main/java/com/jik/feature/detail/component/DetailMovieInfoOverview.kt
@@ -1,0 +1,30 @@
+package com.jik.feature.detail.component
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+internal fun Overview(
+    modifier: Modifier = Modifier,
+    overview: String
+) {
+    Column(modifier = modifier) {
+        Text(
+            text = "OVERVIEW",
+            color = Color(0xFF8A8A8A),
+            style = MaterialTheme.typography.bodyLarge
+        )
+        Spacer(modifier = Modifier.height(8.dp))
+        Text(
+            text = overview,
+            style = MaterialTheme.typography.bodyMedium,
+        )
+    }
+}
